+++ src/generator/compute-model-params/compute-create-dto-params.ts
@@ -2,6 +2,8 @@ import slash from 'slash';
 import path from 'node:path';
 import {
   DTO_CREATE_OPTIONAL,
+  DTO_CREATE_HIDDEN,
+  DTO_CREATE_API_RESP,
   DTO_RELATION_CAN_CONNECT_ON_CREATE,
   DTO_RELATION_CAN_CREATE_ON_CREATE,
   DTO_RELATION_CAN_DISCONNECT_ON_UPDATE,
@@ -32,7 +34,7 @@ import {
 
 import type { DMMF } from '@prisma/generator-helper';
 import type { TemplateHelpers } from '../template-helpers';
-import type {
+import {
   Model,
   CreateDtoParams,
   ImportStatementParams,
@@ -54,6 +56,7 @@ export const computeCreateDtoParams = ({
   templateHelpers,
 }: ComputeCreateDtoParamsParam): CreateDtoParams => {
   let hasApiProperty = false;
+  let hasApiRespProperty = false;
   const imports: ImportStatementParams[] = [];
   const apiExtraModels: string[] = [];
   const extraClasses: string[] = [];
@@ -64,7 +67,9 @@ export const computeCreateDtoParams = ({
 
   const fields = model.fields.reduce((result, field) => {
     const { name } = field;
-    const overrides: Partial<DMMF.Field> = {};
+    const overrides: Partial<DMMF.Field> = {
+      createApiResp: false,
+    };
     const decorators: {
       apiProperties?: IApiProperty[];
       classValidators?: IClassValidator[];
@@ -76,6 +81,7 @@ export const computeCreateDtoParams = ({
     )
       field.isReadOnly = false;
 
+    if (isAnnotatedWith(field, DTO_CREATE_HIDDEN)) return result;
     if (isReadOnly(field)) return result;
     if (isRelation(field)) {
       if (!isAnnotatedWithOneOf(field, DTO_RELATION_MODIFIERS_ON_CREATE)) {
@@ -190,6 +196,8 @@ export const computeCreateDtoParams = ({
     }
 
     if (!templateHelpers.config.noDependencies) {
+      overrides.createApiResp = isAnnotatedWith(field, DTO_CREATE_API_RESP);
+      hasApiRespProperty = hasApiRespProperty || overrides.createApiResp;
       decorators.apiProperties = parseApiProperty(field, {
         type: !overrides.type,
       });
@@ -210,10 +218,11 @@ export const computeCreateDtoParams = ({
     return [...result, mapDMMFToParsedField(field, overrides, decorators)];
   }, [] as ParsedField[]);
 
-  if (apiExtraModels.length || hasApiProperty) {
+  if (apiExtraModels.length || hasApiProperty || hasApiRespProperty) {
     const destruct = [];
     if (apiExtraModels.length) destruct.push('ApiExtraModels');
     if (hasApiProperty) destruct.push('ApiProperty');
+    if (hasApiRespProperty) destruct.push('ApiResponseProperty');
     imports.unshift({ from: '@nestjs/swagger', destruct });
   }