+++ src/generator/compute-model-params/compute-plain-dto-params.ts
@@ -1,6 +1,10 @@
 import slash from 'slash';
 import path from 'node:path';
-import { DTO_ENTITY_HIDDEN, DTO_RELATION_INCLUDE_ID } from '../annotations';
+import {
+  DTO_ENTITY_HIDDEN,
+  DTO_RELATION_INCLUDE_ID,
+  DTO_PLAIN_API_RESP,
+} from '../annotations';
 import { isAnnotatedWith, isRelation, isType } from '../field-classifiers';
 import {
   getRelationScalars,
@@ -32,6 +36,7 @@ export const computePlainDtoParams = ({
   templateHelpers,
 }: ComputePlainDtoParamsParam): PlainDtoParams => {
   let hasApiProperty = false;
+  let hasApiRespProperty = false;
   const imports: ImportStatementParams[] = [];
   const apiExtraModels: string[] = [];
 
@@ -43,6 +48,7 @@ export const computePlainDtoParams = ({
     const overrides: Partial<DMMF.Field> = {
       isRequired: true,
       isNullable: !field.isRequired,
+      plainApiResp: false,
     };
     const decorators: { apiProperties?: IApiProperty[] } = {};
 
@@ -93,6 +99,8 @@ export const computePlainDtoParams = ({
     }
 
     if (!templateHelpers.config.noDependencies) {
+      overrides.plainApiResp = isAnnotatedWith(field, DTO_PLAIN_API_RESP);
+      hasApiRespProperty = hasApiRespProperty || overrides.plainApiResp;
       decorators.apiProperties = parseApiProperty(
         { ...field, isRequired: false, isNullable: !field.isRequired },
         { default: false },
@@ -108,10 +116,11 @@ export const computePlainDtoParams = ({
     return [...result, mapDMMFToParsedField(field, overrides, decorators)];
   }, [] as ParsedField[]);
 
-  if (apiExtraModels.length || hasApiProperty) {
+  if (apiExtraModels.length || hasApiProperty || hasApiRespProperty) {
     const destruct = [];
     if (apiExtraModels.length) destruct.push('ApiExtraModels');
     if (hasApiProperty) destruct.push('ApiProperty');
+    if (hasApiRespProperty) destruct.push('ApiResponseProperty');
     imports.unshift({ from: '@nestjs/swagger', destruct });
   }