+++ CHANGELOG.md
@@ -5,6 +5,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [1.17.5] - 2023-04-29
+
+- add annotation `@DtoCreateApiResponse`, `@DtoUpdateApiResponse`, `@DtoPlainApiResponse`
+- add annotation `@DtoCreateHidden`
+- restore annotation `@DtoUpdateHidden`
+- output NestJs pluralized model folders.
+- add `@ApiProperty`.type for `@DtoCastType'
+- support nested parentheses for validators, e.g.: @Matches(/^[a-z\d](?:[a-z\d]|-(?=[a-z\d])){0,38}$/i)
+- add annotation `@CustomValidator`
+- always set `@ApiProperty`.required
+- fix array: @Matches(/xxx/'', { each: true })
+
+
 ## [1.17.4] - 2023-03-04
 
 - fix connect and create class names that get too many suffixes if `dtoSuffix` does not end with `dto` [#17][i17]